api-gateway/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
mock==5.0.1
PyJWT==2.3.0
pytest==7.2.2
redis==4.5.4
requests==2.25.1
